Added new exporters to the TestRunner and reworked the unit tests.
* The unit test structure is now entirely different: lots of code was broken and outdated...
* The TestRunner project has been updated with multiple exporters.
* The TestRunner now doesn't crash on exit.
* Fixed MSVC warnings.
* This PR also adds unit test enabling control via the Projucer (and juce+cmake I'm guessing).
* Removed the Build folder so as to avoid commit noise.

Note: looks like everything works just fine with VS2017!

+++ blueprint/core/blueprint_CanvasView.h
@@ -320,7 +320,7 @@ namespace blueprint
                         const float cornerSize = args.arguments[4];
 
                         graphics->setColour(properties.strokeStyle.colour);
-                        graphics->drawRoundedRectangle(x, y, width, height, cornerSize, properties.lineWidth);
+                        graphics->drawRoundedRectangle(x, y, width, height, cornerSize, (float) properties.lineWidth);
 
                         return juce::var();
                     }
